Better organized package structure, making grime its own package.
